\docType{methods}
\name{getter}
\alias{getter}
\alias{$,PIVW-method}
\title{Applies method $ to different classes}
\usage{
\S4method{$}{PIVW}(x, name)
}
\arguments{
\item{x}{Object.}

\item{name}{Name of slot.}
}
\description{
Enables slots of objects in this package to be referenced easily.
}
\keyword{internal}
