\docType{data}
\name{example}
\alias{By_exp}
\alias{Bx_exp}
\alias{Byse_exp}
\alias{Bxse_exp}
\title{Example data}


\description{
  \code{By_exp}, \code{Bx_exp}, \code{By_exp} and \code{Bx_exp} are numeric vectors of example data for \code{By}, \code{Bx}, \code{Byse} and \code{Bxse} of 1000 IVs, respectively.
}

\keyword{internal}



